Extract RoutineStoreSpy from Test file
